Add UnifiedUploadValidator and update wrappers
